lmb-1556: re arrange the routes for custom forms
