+++ app/components/shared/compact-menu.hbs
@@ -21,7 +21,7 @@
       </AuLink>
     {{/if}}
     {{#if (is-feature-enabled "editable-forms")}}
-      <AuLink @route="custom-forms.overview" role="menuitem">
+      <AuLink @route="custom-forms" role="menuitem">
         <AuIcon @icon="folder" @alignment="left" />
         Eigen gegevens
       </AuLink>
