+++ app/components/shared/main-menu.hbs
@@ -26,7 +26,7 @@
       {{/if}}
       {{#if (is-feature-enabled "editable-forms")}}
         <li class="au-c-list-navigation__item">
-          <AuNavigationLink @route="custom-forms.overview">
+          <AuNavigationLink @route="custom-forms">
             <AuIcon @icon="folder" @alignment="left" />
             Eigen gegevens
           </AuNavigationLink>
