+++ app/components/shared/modules.hbs
@@ -36,8 +36,7 @@
       <:description>Hier vind je een overzicht van de eigen formulieren binnen
         Lokaal Mandatenbeheer</:description>
       <:link>
-        <AuLink @route="custom-forms.overview" @skin="button">Ga naar eigen
-          gegevens</AuLink>
+        <AuLink @route="custom-forms" @skin="button">Ga naar eigen gegevens</AuLink>
       </:link>
     </LoketModuleCard>
   </li>
