+++ app/controllers/custom-forms/index.js
@@ -3,7 +3,7 @@ import Controller from '@ember/controller';
 import { tracked } from '@glimmer/tracking';
 import { service } from '@ember/service';
 
-export default class CustomFormsOverviewController extends Controller {
+export default class CustomFormsIndexController extends Controller {
   @service toaster;
 
   @tracked page = 0;
