+++ app/routes/custom-forms/index.js
@@ -2,7 +2,7 @@ import Route from '@ember/routing/route';
 
 import { service } from '@ember/service';
 
-export default class CustomFormsOverviewRoute extends Route {
+export default class CustomFormsIndexRoute extends Route {
   @service store;
   @service session;
