+++ app/routes/custom-forms/instances.js
@@ -1,0 +1,28 @@
+import Route from '@ember/routing/route';
+
+import { service } from '@ember/service';
+
+export default class CustomFormsInstancesIndexRoute extends Route {
+  @service store;
+  @service session;
+  @service semanticFormRepository;
+  @service customForms;
+
+  beforeModel(transition) {
+    this.session.requireAuthentication(transition, 'login');
+  }
+
+  async model({ id }) {
+    const form = await this.store.findRecord('form', id);
+    const formDefinition =
+      await this.semanticFormRepository.getFormDefinition(id);
+    const usage = await this.customForms.getFormDefinitionUsageCount(id);
+
+    return {
+      form,
+      formDefinition,
+      headerLabels: await this.semanticFormRepository.getHeaderLabels(form.id),
+      usage,
+    };
+  }
+}
