+++ app/routes/custom-forms/instances/new.js
@@ -4,10 +4,9 @@ import { service } from '@ember/service';
 export default class CustomFormsInstancesNewRoute extends Route {
   @service semanticFormRepository;
 
-  async model({ id }) {
-    const formDefinition =
-      await this.semanticFormRepository.getFormDefinition(id);
+  async model() {
+    const parent = this.modelFor('custom-forms.instances');
 
-    return { formDefinition };
+    return { formDefinition: parent.formDefinition };
   }
 }
