+++ app/templates/custom-forms.hbs
@@ -1,2 +1,2 @@
-{{page-title "Eigen gegevens"}}
+{{breadcrumb "Eigen gegevens" route="custom-forms"}}
 {{outlet}}
