+++ app/templates/custom-forms/index.hbs
@@ -1,5 +1,4 @@
 {{page-title "Overzicht"}}
-{{breadcrumb "Overzicht" route="custom-forms.overview"}}
 <AuToolbar
   @size="large"
   class="au-u-padding-bottom-none au-u-margin-bottom"
@@ -50,7 +49,7 @@
     <c.body as |form|>
       <td class="au-u-1-5">
         <LinkTo
-          @route="custom-forms.instances.index"
+          @route="custom-forms.instances"
           @model={{form.id}}
           class="au-c-link au-u-medium"
         >
@@ -58,7 +57,7 @@
         </LinkTo>
       </td>
       <td>
-        {{form.description}}
+        {{form.id}}
       </td>
       <td class="au-u-1-5">{{form.displayCreatedDate}}</td>
       <td class="au-u-1-5">{{form.displayModifiedDate}}</td>
