+++ app/templates/custom-forms/instances.hbs
@@ -1,0 +1,6 @@
+{{breadcrumb
+  this.model.form.name
+  route="custom-forms.instances.index"
+  model=this.model.form.id
+}}
+{{outlet}}
