+++ app/templates/custom-forms/instances/index.hbs
@@ -1,10 +1,3 @@
-{{page-title "Overzicht"}}
-
-{{breadcrumb
-  this.model.form.name
-  route="custom-forms.instances.index"
-  model=this.model.form.id
-}}
 {{outlet}}
 
 <AuMainContainer as |main|>
