+++ app/templates/custom-forms/instances/instance.hbs
@@ -1,4 +1,5 @@
-{{page-title "Instance"}}
+{{breadcrumb "Formulier" route="custom-forms.instances.instance"}}
+
 <div class="au-o-box au-c-body-container au-c-body-container--scroll">
   <EditableForm
     @instanceId={{this.model.instanceId}}
